+++ src/prepare/PrepareUpload.ts
@@ -2,7 +2,7 @@ import { TextureSource } from '../rendering/renderers/shared/texture/sources/Tex
 import { GraphicsContext } from '../scene/graphics/shared/GraphicsContext';
 import { Text } from '../scene/text/Text';
 import { BitmapText } from '../scene/text-bitmap/BitmapText';
-import { HTMLText } from '../scene/text-html/_HtmlText';
+import { HTMLText } from '../scene/text-html/HTMLText';
 import { PrepareQueue } from './PrepareQueue';
 
 import type { FillInstruction, TextureInstruction } from '../scene/graphics/shared/GraphicsContext';
